Added building steps
I will be honest, dear reader. This commit is the result of a massive
cleanup operation, so I apologize for the state of it.

Once, I though I could get away with running _everything_ through the
Ansible playbook, including the steps done in the build_dist.sh file. I
would interleave them in a cool way using async calls. I was wrong.

The way this has to be done is the expected (and boring) way. The
distribution has to be built by running `build_dist.sh`, then the
playbook can be run to deploy it.